fix: 커뮤니티 비로그인 시 접근 가능하게 수정
