+++ src/screens/(Common)/BottomTabs/Community/CommunityPostList.js
@@ -17,8 +17,6 @@ import AlertModal from '@components/modals/AlertModal';
 import FullScreenImageModal from '@components/modals/FullScreenImageModal';
 import Loading from '@components/Loading';
 import communityApi from '@utils/api/communityApi';
-import useUserStore from '@stores/userStore';
-import {showErrorModal} from '@utils/loginModalHub';
 import {toggleFavorite} from '@utils/toggleFavorite';
 import {COLORS} from '@constants/colors';
 import styles from './Community.styles';
@@ -114,22 +112,11 @@ const CommunityPostList = ({category, selectedSort, isActive}) => {
         setHasNext(false);
         console.warn('fetchCommunityPosts 실패:', error);
 
-        const role = useUserStore.getState().userRole;
-        if (role !== 'USER') {
-          showErrorModal({
-            message: '커뮤니티는\n로그인 후 이용할 수 있어요',
-            buttonText2: '취소',
-            buttonText: '로그인하기',
-            onPress: () => navigation.navigate('Login'),
-            onPress2: () => {},
-          });
-        } else {
-          setErrorModal({
-            visible: true,
-            message: '커뮤니티 글을 불러오지 못했어요. 잠시 후 다시 시도해 주세요.',
-            buttonText: '확인',
-          });
-        }
+        setErrorModal({
+          visible: true,
+          message: '커뮤니티 글을 불러오지 못했어요. 잠시 후 다시 시도해 주세요.',
+          buttonText: '확인',
+        });
       } finally {
         if (isRefresh) {
           setIsRefreshing(false);
@@ -140,7 +127,7 @@ const CommunityPostList = ({category, selectedSort, isActive}) => {
         }
       }
     },
-    [category, selectedSort, navigation],
+    [category, selectedSort],
   );
 
   useFocusEffect(
